Remove dependency on ServerSslConfig.
io.quarkus.runtime.configuration.ssl.ServerSslConfig was moved to
the Vert.x-specific modules in Quarkus commit
96955c901f7a6ca3f51f3c319189e37dc26b3f4b, which breaks the existing
GrpcConfig code.  Since GrpcRecorder expected the SSL context to be null
anyway, this patch removes the (currently unused) configuration
reference and restores compatibility with Quarkus 0.23.1.

Also bumps the artifact version to 0.23.1 in accordance with the Quarkus
release.